Improve the test case extractor for testing
