+++ core/src/main/kotlin/com/karumi/kotlinsnapshot/KotlinSnapshot.kt
@@ -1,8 +1,9 @@
 package com.karumi.kotlinsnapshot
 
 import com.karumi.kotlinsnapshot.core.Camera
-import com.karumi.kotlinsnapshot.core.SerializationModule
 import com.karumi.kotlinsnapshot.core.KotlinSerialization
+import com.karumi.kotlinsnapshot.core.SerializationModule
+import com.karumi.kotlinsnapshot.core.TestCaseExtractor
 
 class KotlinSnapshot<in A>(
     serializationModule: SerializationModule<A>,
@@ -15,14 +16,18 @@ class KotlinSnapshot<in A>(
         )
     }
 
-    private val camera = Camera(serializationModule, snapshotsFolder)
+    private val camera = Camera(
+        serializationModule,
+        TestCaseExtractor(),
+        snapshotsFolder
+    )
 
     fun matchWithSnapshot(value: A, snapshotName: String? = null) {
         camera.matchWithSnapshot(value, snapshotName)
     }
 }
 
-private val camera = Camera(KotlinSerialization())
+private val camera = Camera(KotlinSerialization(), TestCaseExtractor())
 
 fun Any.matchWithSnapshot(snapshotName: String? = null) {
     camera.matchWithSnapshot(this, snapshotName)
